pyproject.toml
[tool.poetry]
name = "get-nft-cost"
version = "0.0.1"
description = "SPbPU&WelTec course paper"
authors = ["Yury Borodich <xkxixnxgx@yandex.ru>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "get_nft_cost"}]

[tool.poetry.dependencies]
python = "3.8.5"
fastapi = "0.85.0"
pandas = "1.5.0"
numpy = "1.23.3"
matplotlib = "3.5"
seaborn = "0.12.1"
requests = "2.28.1"
python-dotenv = "0.21.0"


[tool.poetry.group.dev.dependencies]
ipython = "8.5.0"
jupyter = "1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
